split chunk_nodes_cache test
