+++ integration-tests/src/tests/client/features.rs
@@ -3,6 +3,7 @@
 #[cfg(feature = "protocol_feature_account_id_in_function_call_permission")]
 mod account_id_in_function_call_permission;
 mod cap_max_gas_price;
+mod chunk_nodes_cache;
 #[cfg(feature = "protocol_feature_fix_contract_loading_cost")]
 mod fix_contract_loading_cost;
 mod fix_storage_usage;
